packages/api/pyproject.toml
[project]
name = "magnet-api"
version = "0.1.0"
description = "Magnet API — agentic UA creative production backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "celery>=5.3",
    "redis>=5.0",
    "boto3>=1.34",
    "httpx>=0.27",
    "pydantic>=2.6",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
